+++ packages/components-js/tests/e2e/helpers/puppeteer-helper.ts
@@ -127,16 +127,17 @@ export const selectNode = async (page: Page, selector: string): Promise<ElementH
       : '';
   return (
     await page.evaluateHandle(`document.querySelector('${selectorParts[0].trim()}')${shadowRootSelectors}`)
-  ).asElement();
+  ).asElement() as ElementHandle;
 };
 
-export const getShadowRoot = async (element: ElementHandle): Promise<ElementHandle> =>
-  (await element.evaluateHandle((el) => el.shadowRoot)).asElement();
+export const getShadowRoot = async (element: ElementHandle): Promise<ElementHandle<ShadowRoot>> => {
+  return (await element.evaluateHandle((el) => el.shadowRoot)).asElement();
+};
 
 const containsCapitalChar = (key: string): boolean => /[A-Z]/.test(key);
 
-export const getAttribute = async (element: ElementHandle, attribute: string): Promise<string> => {
-  return await element.evaluate((el: HTMLElement, attr: string) => el.getAttribute(attr), attribute);
+export const getAttribute = (element: ElementHandle, attribute: string): Promise<string> => {
+  return element.evaluate((el: HTMLElement, attr: string) => el.getAttribute(attr), attribute);
 };
 
 export const setAttribute = async (element: ElementHandle, key: string, value: string): Promise<void> => {
@@ -169,19 +170,19 @@ export const getCssClasses = async (element: ElementHandle): Promise<string> => 
   return Object.values(await getProperty(element, 'classList')).join(' ');
 };
 
-export const getActiveElementTagNameInShadowRoot = async (element: ElementHandle): Promise<string> => {
+export const getActiveElementTagNameInShadowRoot = (element: ElementHandle): Promise<string> => {
   return element.evaluate((el) => el.shadowRoot.activeElement.tagName);
 };
 
-export const getActiveElementClassNameInShadowRoot = async (element: ElementHandle): Promise<string> => {
+export const getActiveElementClassNameInShadowRoot = (element: ElementHandle): Promise<string> => {
   return element.evaluate((el) => el.shadowRoot.activeElement.className);
 };
 
-export const getActiveElementId = async (page: Page): Promise<string> => {
+export const getActiveElementId = (page: Page): Promise<string> => {
   return page.evaluate(() => document.activeElement.id);
 };
 
-export const getActiveElementTagName = async (page: Page): Promise<string> => {
+export const getActiveElementTagName = (page: Page): Promise<string> => {
   return page.evaluate(() => document.activeElement.tagName);
 };
 
@@ -238,12 +239,11 @@ export const getElementPositions = (
   }, element);
 };
 
-export const reattachElement = async (page: Page, selector: string): Promise<void> => {
-  await page.evaluate((selector: string) => {
-    const [element] = Array.from(document.getElementsByTagName(selector));
-    element.remove();
-    document.body.appendChild(element);
-  }, selector);
+export const reattachElementHandle = (handle: ElementHandle): Promise<void> => {
+  return handle.evaluate((el) => {
+    el.remove();
+    document.body.appendChild(el);
+  });
 };
 
 export const enableBrowserLogging = (page: Page): void => {
@@ -273,8 +273,19 @@ export const initConsoleObserver = (page: Page): void => {
     }
   });
 };
-export const getConsoleErrorsAmount = () => consoleMessages.filter((x) => x.type() === 'error').length;
-export const getConsoleWarningsAmount = () => consoleMessages.filter((x) => x.type() === 'warning').length;
+
+const getConsoleErrors = () => consoleMessages.filter((x) => x.type() === 'error');
+const getConsoleWarnings = () => consoleMessages.filter((x) => x.type() === 'warning');
+export const getConsoleErrorsAmount = () => getConsoleErrors().length;
+export const getConsoleErrorMessages = () =>
+  getConsoleErrors()
+    .map((msg) => '- ' + msg.text())
+    .join('\n');
+export const getConsoleWarningsAmount = () => getConsoleWarnings().length;
+export const getConsoleWarningMessages = () =>
+  getConsoleWarnings()
+    .map((msg) => '- ' + msg.text())
+    .join('\n');
 
 const thrownErrors: string[] = [];
 
@@ -348,7 +359,7 @@ export const expectA11yToMatchSnapshot = async (
     ...options,
   });
 
-  message ? expect(snapshot, message).toMatchSnapshot(message) : expect(snapshot).toMatchSnapshot();
+  message ? expect(snapshot).toMatchSnapshot(message) : expect(snapshot).toMatchSnapshot();
 };
 
 export const expectToSkipFocusOnComponent = async (page: Page, component: ElementHandle, before: ElementHandle) => {