+++ packages/components-js/tests/vrt/playwright/specs/overview.vrt.ts
@@ -1,0 +1,6 @@
+import { test } from '@playwright/test';
+import { executeVisualRegressionTest, openPopovers } from '../helpers/playwright-helper';
+
+test.describe('should have no visual regression', async () => {
+  await executeVisualRegressionTest('overview', { scenario: (page) => openPopovers(page) });
+});