+++ packages/components/src/components/common/scroller/scroller-utils.spec.ts
@@ -1,7 +1,6 @@
 import type { Direction } from './scroller-utils';
-import { getScrollerElements, getScrollPositionAfterPrevNextClick } from './scroller-utils';
-import * as getShadowRootHTMLElementUtil from '../../../utils/dom/getShadowRootHTMLElement';
-import * as getHTMLElementUtil from '../../../utils/dom/getHTMLElement';
+import { getScrollerElements, getScrollPositionAfterPrevNextClick, isScrollable } from './scroller-utils';
+import * as getHTMLElementsUtils from '../../../utils/dom/getHTMLElements';
 
 describe('getScrollPositionAfterPrevNextClick()', () => {
   it.each<[number, number, Direction, number]>([
@@ -19,18 +18,36 @@ describe('getScrollPositionAfterPrevNextClick()', () => {
 });
 
 describe('getScrollerElements()', () => {
-  it('should call getShadowRootHTMLElement with correct selectors', () => {
-    const spy = jest.spyOn(getShadowRootHTMLElementUtil, 'getShadowRootHTMLElement').mockImplementationOnce(() => {
-      return {} as Element;
-    });
-    jest.spyOn(getHTMLElementUtil, 'getHTMLElement').mockImplementationOnce(() => {
-      return {} as Element;
-    });
+  it('should call getHTMLElements() with correct parameters', () => {
+    const spy = jest.spyOn(getHTMLElementsUtils, 'getHTMLElements');
+    const scroller = document.createElement('p-scroller');
+    scroller.attachShadow({ mode: 'open' });
+
+    getScrollerElements(scroller);
+    expect(spy).toBeCalledWith(scroller.shadowRoot, '.scroll-area,.action-prev');
+  });
+
+  it('should return tuple result of getHTMLElements()', () => {
+    const mockResult1 = document.createElement('div');
+    mockResult1.id = 'mock-result-1';
+    const mockResult2 = document.createElement('div');
+    mockResult2.id = 'mock-result-2';
+    jest.spyOn(getHTMLElementsUtils, 'getHTMLElements').mockReturnValue([mockResult1, mockResult2]);
 
     const scroller = document.createElement('p-scroller');
-    getScrollerElements(scroller);
+    scroller.attachShadow({ mode: 'open' });
 
-    expect(spy).toBeCalledWith(scroller, '.scroll-area');
-    expect(spy).toBeCalledWith(scroller, '.gradient');
+    expect(getScrollerElements(scroller)).toEqual([mockResult1, mockResult2]);
+  });
+});
+
+describe('isScrollable()', () => {
+  it.each<[{ isPrevHidden: boolean; isNextHidden: boolean }, boolean]>([
+    [{ isPrevHidden: true, isNextHidden: true }, false],
+    [{ isPrevHidden: false, isNextHidden: false }, true],
+    [{ isPrevHidden: false, isNextHidden: true }, true],
+    [{ isPrevHidden: true, isNextHidden: false }, true],
+  ])('should for %o return %s', ({ isPrevHidden, isNextHidden }, expected) => {
+    expect(isScrollable(isPrevHidden, isNextHidden)).toBe(expected);
   });
 });