+++ packages/storefront/src/utils/index.ts
@@ -1,4 +1,5 @@
 export * from './code-highlighting';
-export * from './getAnchorLink';
 export * from './formatting';
+export * from './getAnchorLink';
+export * from './stackblitz';
 export * from './validation';